Steady State through independent critical_cut and touchdown_touch configurations.
- integrated new steady state evaluation method.
- disessembled criteria_evaluator to split envelopes from CC and SS logic
- modified analyzer rasterize_logic to use denser window around boundary segments
- modified CC logic: damping in place, passing max iterations, passing min_force_results and removing redundant rasterize calls by passing x_coords into find_crack_length
- added tests for new rasterize function
- removed numpy warnings by casting to floats from arrays when returning.
