+++ CITATION.cff
@@ -8,7 +8,7 @@ authors:
 - family-names: "Weissgraeber"
   given-names: "Philipp"
   orcid: "https://orcid.org/0000-0001-8320-8672"
-version: 3.1.9
+version: 3.2.0
 date-released: 2021-12-30
 identifiers:
   - description: Collection of archived snapshots of all versions of WEAC
