+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "weac"
-version = "3.1.9"
+version = "3.2.0"
 authors = [{ name = "2phi GbR", email = "mail@2phi.de" }]
 description = "Weak layer anticrack nucleation model"
 readme = "README.md"
@@ -126,7 +126,7 @@ ignore = [
 ]
 
 [tool.bumpversion]
-current_version = "3.1.9"
+current_version = "3.2.0"
 
 [[tool.bumpversion.files]]
 filename = "pyproject.toml"
