+++ src/weac/__init__.py
@@ -2,4 +2,4 @@
 WEAC - Weak Layer Anticrack Nucleation Model
 """
 
-__version__ = "3.1.9"
+__version__ = "3.2.0"
