+++ src/weac/analysis/__init__.py
@@ -3,14 +3,19 @@ This package contains modules for analyzing the results of the WEAC model.
 """
 
 from .analyzer import Analyzer
-from .criteria_evaluator import (
+from .coupled_criterion import (
     CoupledCriterionHistory,
     CoupledCriterionResult,
-    CriteriaEvaluator,
     FindMinimumForceResult,
-    SteadyStateResult,
+    MaximalStressResult,
 )
+from .criteria_evaluator import CriteriaEvaluator
 from .plotter import Plotter
+from .steady_state import (
+    SteadyStateErrBlock,
+    SteadyStateResult,
+    SteadyStateTensileBlock,
+)
 
 __all__ = [
     "Analyzer",
@@ -18,6 +23,9 @@ __all__ = [
     "CoupledCriterionHistory",
     "CoupledCriterionResult",
     "FindMinimumForceResult",
+    "MaximalStressResult",
+    "SteadyStateErrBlock",
     "SteadyStateResult",
+    "SteadyStateTensileBlock",
     "Plotter",
 ]
